refactor: rename embedded script hotfix package
Clarify the ANC-owned script hotfix applier naming and document the 15 supported-version compatibility rotation.


Copilot-Session: 866474a8-0bd8-4410-9076-4e37e0e3bc81
