+++ aks-node-controller/scripthotfix/applier_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
 
-package hotfixpayload
+package scripthotfix
 
 import (
 	"encoding/json"
